+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  * Added ad-rss type-stubs for the PythonAPI when building with RSS support
  * Make TrafficManager PID controller use actual delta times to improve robustness to different fixed_delta_seconds
  * ROS2Native: Extended functionality and performance of ROS2 support
+ * ROS2Native: Update to fastdds 2.14.6
  * Added Ubuntu 24.04 support
  
 ## CARLA 0.9.16